requirements.txt
anthropic>=0.18.0
elevenlabs>=0.2.0
google-generativeai>=0.3.0
lxml>=4.9.0
openai>=1.12.0
pandas>=2.0.0
pillow>=10.0.0
praw>=7.0.0
pydantic>=2.0.0
PyPDF2>=3.0.0
pydub>=0.25.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
requests>=2.31.0
rich>=13.0.0
speechmatics>=1.8.0
tiktoken>=0.5.0